add lock files in repo

+++ fabfile.py
@@ -138,23 +138,23 @@ def upload_lock_files():
 
 
 def git_remove_lock_and_styles():
-    try:
-        local('git rm --cached *installed.lock')
-        local('git rm --cached *variables.scss')
-        # sed("/home/popov/ac_template_site/.gitignore", "_variables.scss", "# _variables.scss")
-        # sed(CWD+'.gitignore', "installed.lock", "# installed.lock")
-        # local("sed 's/PROJECT_NAME/{}/g; \
-        #             s/DOMAIN_NAME/{}/g; \
-        #             s/USERNAME/{}/g' \
-        #         nginx_config_template > {}_nginx".format(
-        #     env.project_name, env.domain_name, env.user, env.project_name))
-        local("sed -i 's/# installed.lock/installed.lock/g; \
-            s/# _variables.scss/_variables.scss/g' .gitignore")
-        local('git add .')
-        local('git commit -m "remove lock files and scss variables from repo before update"')
-        # .gitignore
-    except Exception as e:
-        print('ERROR REMOVING LOCK FILES', e)
+    output = local("git ls-files *.lock", capture=True)+local("git ls-files *variables.scss", capture=True)
+    if len(output) == 0:
+        return
+    local('git rm --cached *installed.lock')
+    local('git rm --cached *variables.scss')
+    # sed("/home/popov/ac_template_site/.gitignore", "_variables.scss", "# _variables.scss")
+    # sed(CWD+'.gitignore', "installed.lock", "# installed.lock")
+    # local("sed 's/PROJECT_NAME/{}/g; \
+    #             s/DOMAIN_NAME/{}/g; \
+    #             s/USERNAME/{}/g' \
+    #         nginx_config_template > {}_nginx".format(
+    #     env.project_name, env.domain_name, env.user, env.project_name))
+    local("sed -i 's/installed.lock/# installed.lock/g; \
+        s/_variables.scss/# _variables.scss/g' .gitignore")
+    local('git add .')
+    local('git commit -m "remove lock files and scss variables from repo before update"')
+    # .gitignore
 
 
 def git_add_lock_files_and_styles():
